Expand locomotion controller
