+++ Assets/Scripts/ActiveRagdoll/RagdollController.cs
@@ -8,24 +8,15 @@ using Utils;
 public class RagdollController : MonoBehaviour
 {
     [SerializeField] private RagdollJointHandler jointHandler;
-    [SerializeField] private Transform centerOfMass;
     [SerializeField] private RagdollImpactHandler impactHandler;
-    [Header("Movement Properties")] public bool forwardIsCameraDirection = true;
-    public float moveSpeed = 10f;
-    public float turnSpeed = 6f;
-    public float jumpForce = 18f;
-
-    public float StepDuration = 0.2f;
-    public float StepHeight = 1.7f;
-    public float FeetMountForce = 25f;
-
+    [SerializeField] private RagdollLocomotionController locomotionController;
+    [SerializeField] private bool forwardIsCameraDirection = true;
+    [SerializeField] private Transform centerOfMass;
+    [SerializeField] private float turnSpeed = 6f;
     [Header("Reach Properties")] public float reachSensitivity = 25f;
-
     [Header("Actions")] public bool canPunch = true;
     public float punchForce = 15f;
-
     private readonly RagdollInputHandler inputHandler = new();
-    private RagdollLocomotionController locomotionController;
     private readonly RagdollState ragdollState = new();
     [SerializeField] private Camera cam;
     private Vector3 CenterOfMassPoint; //TODO: Check usage
@@ -37,8 +28,8 @@ public class RagdollController : MonoBehaviour
     private void Awake()
     {
         inputHandler.Init();
-        locomotionController = new RagdollLocomotionController(jointHandler, ragdollState);
         defaultTargetState = new RagdollDefaultTargetState(jointHandler);
+        locomotionController.Init(jointHandler, ragdollState, defaultTargetState);
         SetupHandContacts();
         SetupFeetContacts();
         SetupRagdollImpactContacts();
@@ -75,7 +66,7 @@ public class RagdollController : MonoBehaviour
     {
         if (!ragdollState.inAir)
         {
-            PlayerMovement();
+            locomotionController.PlayerMovement(forwardIsCameraDirection, inputHandler.MovementAxis);
 
             if (canPunch)
             {
@@ -93,10 +84,10 @@ public class RagdollController : MonoBehaviour
 
     private void FixedUpdate()
     {
-        PerformWalking();
+        locomotionController.PerformWalking();
         PerformPlayerRotation();
         ResetPlayerPose();
-        PerformPlayerGetUpJumping();
+        locomotionController.PerformPlayerGetUpJumping(inputHandler.JumpValue);
     }
 
     private void PerformPlayerRotation()
@@ -172,177 +163,6 @@ public class RagdollController : MonoBehaviour
     private bool CanResetPoseAfterLanding()
     {
         return ragdollState.inAir && !ragdollState.isJumping && !ragdollState.jumping;
-    }
-
-    private void PerformPlayerGetUpJumping()
-    {
-        if (inputHandler.JumpValue > 0)
-        {
-            if (!ragdollState.jumpAxisUsed)
-            {
-                if (locomotionController.balanced && !ragdollState.inAir)
-                {
-                    ragdollState.jumping = true;
-                }
-
-                else if (!locomotionController.balanced)
-                {
-                    locomotionController.DeactivateRagdoll();
-                }
-            }
-
-            ragdollState.jumpAxisUsed = true;
-        }
-
-        else
-        {
-            ragdollState.jumpAxisUsed = false;
-        }
-
-
-        if (ragdollState.jumping)
-        {
-            ragdollState.isJumping = true;
-
-            Rigidbody rootRigidbody = jointHandler.GetRigidBodyFromJoint(RagdollParts.ROOT);
-            rootRigidbody.velocity = rootRigidbody.velocity.ModifyY(rootRigidbody.transform.up.y * jumpForce);
-        }
-
-        if (ragdollState.isJumping)
-        {
-            ragdollState.jumpingResetTimer += Time.fixedDeltaTime;
-
-            if (ragdollState.jumpingResetTimer > 0.2f)
-            {
-                ragdollState.jumpingResetTimer = 0.0f;
-                ragdollState.jumping = false;
-                ragdollState.isJumping = false;
-                ragdollState.inAir = true;
-            }
-        }
-    }
-
-    private void PlayerMovement()
-    {
-        if (forwardIsCameraDirection)
-        {
-            MoveInCameraDirection();
-        }
-        else
-        {
-            MoveInOwnDirection();
-        }
-    }
-
-    private void MoveInCameraDirection()
-    {
-        var direction = jointHandler.GetRagdollJointWithID(RagdollParts.ROOT).transform.rotation *
-                        new Vector3(inputHandler.MovementAxis.x, 0.0f, inputHandler.MovementAxis.y);
-        direction.y = 0f;
-        Rigidbody rootRigidbody = jointHandler.GetRigidBodyFromJoint(RagdollParts.ROOT);
-        var velocity = rootRigidbody.velocity;
-        rootRigidbody.velocity = Vector3.Lerp(velocity,
-            (direction * moveSpeed) + new Vector3(0, velocity.y, 0), 0.8f);
-
-        if (inputHandler.MovementAxis.x != 0 || inputHandler.MovementAxis.y != 0 && locomotionController.balanced)
-        {
-            StartWalkingForward();
-        }
-        else if (inputHandler.MovementAxis is { x: 0, y: 0 })
-        {
-            StopWalkingForward();
-        }
-    }
-
-    private void StartWalkingForward()
-    {
-        if (!locomotionController.walkForward && !ragdollState.moveAxisUsed)
-        {
-            locomotionController.walkForward = true;
-            ragdollState.moveAxisUsed = true;
-            ragdollState.isKeyDown = true;
-        }
-    }
-
-    private void StopWalkingForward()
-    {
-        if (locomotionController.walkForward && ragdollState.moveAxisUsed)
-        {
-            locomotionController.walkForward = false;
-            ragdollState.moveAxisUsed = false;
-            ragdollState.isKeyDown = false;
-        }
-    }
-
-    private void MoveInOwnDirection()
-    {
-        if (inputHandler.MovementAxis.y != 0)
-        {
-            var rootRigidbody = jointHandler.GetRigidBodyFromJoint(RagdollParts.ROOT);
-            var v3 = rootRigidbody.transform.forward * (inputHandler.MovementAxis.y * moveSpeed);
-            v3.y = rootRigidbody.velocity.y;
-            rootRigidbody.velocity = v3;
-        }
-
-        if (inputHandler.MovementAxis.y > 0)
-        {
-            StartWalkingForwardInOwnDirection();
-        }
-        else if (inputHandler.MovementAxis.y < 0)
-        {
-            StartWalkingBackward();
-        }
-        else
-        {
-            StopWalking();
-        }
-    }
-
-    private void StartWalkingForwardInOwnDirection() =>
-        SetWalkMovingState(() => (!locomotionController.walkForward && !ragdollState.moveAxisUsed), true, false,
-            true, true,
-            jointHandler.PoseOn);
-
-    private void StartWalkingBackward() =>
-        SetWalkMovingState(() => !locomotionController.walkBackward && !ragdollState.moveAxisUsed, false, true,
-            true, true,
-            jointHandler.PoseOn);
-
-    private void StopWalking() => SetWalkMovingState(
-        () => locomotionController.walkForward ||
-              locomotionController.walkBackward && ragdollState.moveAxisUsed, false, false,
-        false, false, jointHandler.DriveOff);
-
-    private void SetWalkMovingState(Func<bool> activationCondition, bool walkForwardSetState, bool walkBackwardSetState,
-        bool isMoveAxisUsed, bool isKeyCurrentlyDown, in JointDrive legsJointDrive)
-    {
-        if (activationCondition.Invoke())
-        {
-            InternalChangeWalkState(walkForwardSetState, walkBackwardSetState, isMoveAxisUsed, isKeyCurrentlyDown,
-                in legsJointDrive);
-        }
-    }
-
-    private void InternalChangeWalkState(bool walkForward, bool walkBackward, bool isMoveAxisUsed,
-        bool isKeyCurrentlyDown,
-        in JointDrive legsJointDrive)
-    {
-        locomotionController.walkForward = walkForward;
-        locomotionController.walkBackward = walkBackward;
-        ragdollState.moveAxisUsed = isMoveAxisUsed;
-        ragdollState.isKeyDown = isKeyCurrentlyDown;
-        if (locomotionController.isRagdoll)
-            SetJointAngularDrivesForLegs(in legsJointDrive);
-    }
-
-    private void SetJointAngularDrivesForLegs(in JointDrive jointDrive)
-    {
-        jointHandler.SetJointAngularDrives(RagdollParts.UPPER_RIGHT_LEG, in jointDrive);
-        jointHandler.SetJointAngularDrives(RagdollParts.LOWER_RIGHT_LEG, in jointDrive);
-        jointHandler.SetJointAngularDrives(RagdollParts.UPPER_LEFT_LEG, in jointDrive);
-        jointHandler.SetJointAngularDrives(RagdollParts.LOWER_LEFT_LEG, in jointDrive);
-        jointHandler.SetJointAngularDrives(RagdollParts.RIGHT_FOOT, in jointDrive);
-        jointHandler.SetJointAngularDrives(RagdollParts.LEFT_FOOT, in jointDrive);
     }
 
     private void PlayerReach()
@@ -477,146 +297,6 @@ public class RagdollController : MonoBehaviour
         RagdollParts.UPPER_RIGHT_ARM,
         RagdollParts.LOWER_RIGHT_ARM, RightHandRigidBody, () => defaultTargetState.UpperRightArmTarget,
         () => defaultTargetState.LowerLeftArmTarget);
-
-    private void PerformWalking()
-    {
-        if (ragdollState.inAir)
-            return;
-
-        if (locomotionController.walkForward)
-        {
-            WalkForwards();
-        }
-
-        if (locomotionController.walkBackward)
-        {
-            WalkBackwards();
-        }
-
-        if (locomotionController.stepRight)
-        {
-            TakeStepRight();
-        }
-        else
-        {
-            ResetStepRight();
-        }
-
-        if (locomotionController.stepLeft)
-        {
-            TakeStepLeft();
-        }
-        else
-        {
-            ResetStepLeft();
-        }
-    }
-
-    private void ResetStepLeft() =>
-        ResetStep(RagdollParts.UPPER_LEFT_LEG, RagdollParts.LOWER_LEFT_LEG, defaultTargetState.UpperLeftLegTarget,
-            defaultTargetState.LowerLeftLegTarget, 7f, 18f);
-
-    private void ResetStepRight() => ResetStep(RagdollParts.UPPER_RIGHT_LEG, RagdollParts.LOWER_RIGHT_LEG,
-        defaultTargetState.UpperRightLegTarget,
-        defaultTargetState.LowerRightLegTarget, 8f, 17f);
-
-    private void ResetStep(string upperLegLabel,
-        string lowerLegLabel,
-        Quaternion upperLegTarget,
-        Quaternion lowerLegTarget,
-        float upperLegLerpMultiplier,
-        float lowerLegLerpMultiplier)
-    {
-        jointHandler.GetConfigurableJointWithID(upperLegLabel).targetRotation = Quaternion.Lerp(
-            jointHandler.GetJointTargetRotation(upperLegLabel), upperLegTarget,
-            upperLegLerpMultiplier * Time.fixedDeltaTime);
-        jointHandler.GetConfigurableJointWithID(lowerLegLabel).targetRotation = Quaternion.Lerp(
-            jointHandler.GetJointTargetRotation(lowerLegLabel), lowerLegTarget,
-            lowerLegLerpMultiplier * Time.fixedDeltaTime);
-
-        Vector3 feetForce = -Vector3.up * (FeetMountForce * Time.deltaTime);
-        jointHandler.GetRigidBodyFromJoint(RagdollParts.RIGHT_FOOT).AddForce(feetForce, ForceMode.Impulse);
-        jointHandler.GetRigidBodyFromJoint(RagdollParts.LEFT_FOOT).AddForce(feetForce, ForceMode.Impulse);
-    }
-
-    private void TakeStepLeft() => TakeStep(ref locomotionController.stepLTimer, RagdollParts.LEFT_FOOT,
-        ref locomotionController.stepLeft, ref locomotionController.stepRight,
-        RagdollParts.UPPER_LEFT_LEG,
-        RagdollParts.LOWER_LEFT_LEG, RagdollParts.UPPER_RIGHT_LEG);
-
-    private void TakeStepRight() => TakeStep(ref locomotionController.stepRTimer, RagdollParts.RIGHT_FOOT,
-        ref locomotionController.stepRight, ref locomotionController.stepLeft,
-        RagdollParts.UPPER_RIGHT_LEG,
-        RagdollParts.LOWER_RIGHT_LEG, RagdollParts.UPPER_LEFT_LEG);
-
-    private void TakeStep(ref float stepTimer,
-        string footLabel,
-        ref bool stepFootState,
-        ref bool oppositeStepFootState,
-        string upperJointLabel,
-        string lowerJointLabel,
-        string upperOppositeJointLabel)
-    {
-        stepTimer += Time.fixedDeltaTime;
-        jointHandler.GetRigidBodyFromJoint(footLabel)
-            .AddForce(-Vector3.up * (FeetMountForce * Time.deltaTime), ForceMode.Impulse);
-
-        var upperLegJoint = jointHandler.GetConfigurableJointWithID(upperJointLabel);
-        var upperLegJointTargetRotation = upperLegJoint.targetRotation;
-
-        var lowerLegJoint = jointHandler.GetConfigurableJointWithID(lowerJointLabel);
-        var lowerLegJointTargetRotation = lowerLegJoint.targetRotation;
-
-        var upperOppositeLegJoint = jointHandler.GetConfigurableJointWithID(upperOppositeJointLabel);
-        var upperOppositeLegJointTargetRotation = upperOppositeLegJoint.targetRotation;
-
-        bool isWalking = locomotionController.walkForward || locomotionController.walkBackward;
-
-        if (locomotionController.walkForward)
-        {
-            upperLegJointTargetRotation = upperLegJointTargetRotation.DisplaceX(0.09f * StepHeight);
-            lowerLegJointTargetRotation = lowerLegJointTargetRotation.DisplaceX(-0.09f * StepHeight * 2);
-            upperOppositeLegJointTargetRotation =
-                upperOppositeLegJointTargetRotation.DisplaceX(-0.12f * StepHeight / 2);
-        }
-
-        if (locomotionController.walkBackward)
-        {
-            //TODO: Is this necessary for something? It's multiplying by 0.
-            upperLegJointTargetRotation = upperLegJointTargetRotation.DisplaceX(-0.00f * StepHeight);
-            lowerLegJointTargetRotation = lowerLegJointTargetRotation.DisplaceX(-0.07f * StepHeight * 2);
-            upperOppositeLegJointTargetRotation = upperOppositeLegJointTargetRotation.DisplaceX(0.02f * StepHeight / 2);
-        }
-
-        if (isWalking)
-        {
-            upperLegJoint.targetRotation = upperLegJointTargetRotation;
-            lowerLegJoint.targetRotation = lowerLegJointTargetRotation;
-            upperOppositeLegJoint.targetRotation = upperOppositeLegJointTargetRotation;
-        }
-
-
-        if (stepTimer <= StepDuration)
-            return;
-
-        stepTimer = 0;
-        stepFootState = false;
-
-        if (isWalking)
-        {
-            oppositeStepFootState = true;
-        }
-    }
-
-    private void WalkBackwards() => locomotionController.Walk(RagdollParts.LEFT_FOOT, RagdollParts.RIGHT_FOOT,
-        ref locomotionController.stepLeft, ref locomotionController.stepRight,
-        ref locomotionController.alertLegLeft,
-        ref locomotionController.alertLegRight);
-
-    private void WalkForwards() => locomotionController.Walk(RagdollParts.RIGHT_FOOT, RagdollParts.LEFT_FOOT,
-        ref locomotionController.stepRight, ref locomotionController.stepLeft,
-        ref locomotionController.alertLegRight,
-        ref locomotionController.alertLegLeft);
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     private void UpdateCenterOfMass()
